refactor: deduplicate dashboard mutation table components
Pull the shared bits of the four mutation tables (pagination, loading
skeletons, the moment/amount columns, amount coloring, and the
mutation/inactive-cost modal wiring) into one MutationsTable.vue.

MutationsBalance, MutationsPOS, MutationsSeller, and MutationsTransfers
now just supply their own extra columns through a scoped slot and
adapt their fetch call (getMutations/getTransfers) to the shared
fetchPage prop. Every wrapper keeps its original props, so call sites
didn't need touching.

Closes #847
